requirements.txt
# Django Core
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
drf-spectacular==0.27.1

# Database
psycopg2-binary==2.9.9

# WebSocket Support
channels==4.0.0
channels-redis==4.2.0
daphne==4.1.0

# Environment Variables
python-decouple==3.8

# Validation
django-phonenumber-field==7.3.0
phonenumbers==8.13.27

# Testing
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-asyncio==0.23.3
factory-boy==3.3.0

# Code Quality
flake8==7.0.0
black==24.1.0
mypy==1.8.0
django-stubs==4.2.7

# Redis
redis==5.0.1

# Utilities
django-extensions==3.2.3
